+++ CHANGELOG.md
@@ -13,6 +13,12 @@ You can find and compare releases at the [GitHub release page](https://github.co
 
 - Add `columns` and `distinct` args to `@count` directive https://github.com/nuwave/lighthouse/pull/2138
 
+## v5.50.0
+
+### Added
+
+- Add `@cacheControl` directive https://github.com/nuwave/lighthouse/pull/2136
+
 ## v5.49.0
 
 ### Added